Move server config functions out of api10.go
Make room for other functions that will be added later.
